fileio.c printf x86_64 compile warning fixes
git-svn-id: https://c64preservation.com/svn/nibtools/trunk@396 048eaad1-c872-4173-8c30-523bb29f53ec

+++ fileio.c
@@ -473,7 +473,7 @@ int read_nb2(char *filename, BYTE *track_buffer, BYTE *track_density, size_t *tr
 		if(track_density[track] & BM_NO_SYNC) printf("NOSYNC!");
 		if(track_density[track] & BM_FF_TRACK) printf("KILLER!");
 
-		printf("%d:%d) (pass %zu, %zu errors) %.1lu%%",
+		printf("%d:%zu) (pass %zu, %zu errors) %.1lu%%",
 			track_density[track]&3, track_length[track],
 			best_pass, best_err,
 			((track_length[track] / capacity[track_density[track]&3]) * 100));
@@ -566,8 +566,9 @@ int read_g64(char *filename, BYTE *track_buffer, BYTE *track_density, size_t *tr
 			if(track_density[track] & BM_NO_SYNC) printf("NOSYNC!");
 			if(track_density[track] & BM_FF_TRACK) printf("KILLER!");
 
-			printf("%d:%d) %.1f%%\n", track_density[track]&3, track_length[track],
-				((float)track_length[track] / (float)capacity[track_density[track]&3]) * 100);
+			printf("%d:%zu) %.1lu%%\n",
+				track_density[track]&3, track_length[track],
+				((track_length[track] / capacity[track_density[track]&3]) * 100));
 		}
 	}
 	fclose(fpin);
